Dry it off and make tooltips look more like wiki pages
